refactor: 인수 테스트에서 사용하는 객체 패키지 변경
