+++ src/test/java/kitchenpos/acceptance/MenuAcceptanceTest.java
@@ -5,9 +5,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.List;
-import kitchenpos.acceptance.common.MenuGroupHttpCommunication;
-import kitchenpos.acceptance.common.MenuHttpCommunication;
-import kitchenpos.acceptance.common.ProductHttpCommunication;
+import kitchenpos.acceptance.common.httpcommunication.MenuGroupHttpCommunication;
+import kitchenpos.acceptance.common.httpcommunication.MenuHttpCommunication;
+import kitchenpos.acceptance.common.httpcommunication.ProductHttpCommunication;
 import kitchenpos.acceptance.common.fixture.RequestBody;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
