+++ src/test/java/kitchenpos/acceptance/MenuGroupAcceptanceTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.List;
-import kitchenpos.acceptance.common.MenuGroupHttpCommunication;
+import kitchenpos.acceptance.common.httpcommunication.MenuGroupHttpCommunication;
 import kitchenpos.acceptance.common.fixture.RequestBody;
 import kitchenpos.domain.MenuGroup;
 import org.junit.jupiter.api.DisplayName;
