+++ src/test/java/kitchenpos/acceptance/OrderAcceptanceTest.java
@@ -6,19 +6,18 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.List;
 import java.util.Map;
-import kitchenpos.acceptance.common.HttpCommunication;
-import kitchenpos.acceptance.common.MenuGroupHttpCommunication;
-import kitchenpos.acceptance.common.MenuHttpCommunication;
-import kitchenpos.acceptance.common.OrderHttpCommunication;
-import kitchenpos.acceptance.common.OrderTableHttpCommunication;
-import kitchenpos.acceptance.common.ProductHttpCommunication;
+import kitchenpos.acceptance.common.httpcommunication.HttpCommunication;
+import kitchenpos.acceptance.common.httpcommunication.MenuGroupHttpCommunication;
+import kitchenpos.acceptance.common.httpcommunication.MenuHttpCommunication;
+import kitchenpos.acceptance.common.httpcommunication.OrderHttpCommunication;
+import kitchenpos.acceptance.common.httpcommunication.OrderTableHttpCommunication;
+import kitchenpos.acceptance.common.httpcommunication.ProductHttpCommunication;
 import kitchenpos.acceptance.common.fixture.RequestBody;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.Product;
-import kitchenpos.ui.OrderRestController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
