+++ src/test/java/kitchenpos/acceptance/common/httpcommunication/OrderTableHttpCommunication.java
@@ -1,4 +1,4 @@
-package kitchenpos.acceptance.common;
+package kitchenpos.acceptance.common.httpcommunication;
 
 import java.util.Map;
